+++ test/src/types/z_string_test.dart
@@ -24,7 +24,7 @@ void main() {
             validInputs: baseValidInputs,
             invalidInputs: [
               ...baseInvalidInputs,
-              (input: null, expected: const [ZIssueParseFail(from: Null, to: String, val: null)]),
+              (input: null, expected: const [ZIssueRequired()]),
             ],
           ),
           ZString(),